changed the create table query to int

+++ createtablesASPire.sql
@@ -303,7 +303,7 @@ DROP TABLE IF EXISTS "config_marine_sensors";
 CREATE TABLE config_marine_sensors (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
   loop_time DOUBLE,
-  time_interval DOUBLE,
+  time_interval INTEGER,
   measure_at_checkpoint BOOLEAN
 );
